requirements.txt
azure-core==1.30.1
azure-storage-blob==12.20.0
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
idna==3.7
isodate==0.6.1
numpy==1.26.4
pandas==2.2.2
paramiko==3.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
